Replaced unsigned int unpack template to used signed (per puzzle.h).
RT-113716: signature_as_char_string was using the unsigned unpack
template to convert the signature to a character representation; the
reporter pointed out that the signature is actually a signed character
and that the resulting string representation was not correct.

https://rt.cpan.org/Ticket/Display.html?id=113716

The suggested fix was implemented; the unpack now uses the signed char
template (c); this results in an array of values that range -2..2; the
arrays is then iterated and 2 is added to each value; this shifts the
range to 0..4 (removes the minus sign), and therefore produces stringified
signatures that are of a consistent length from image to image. Now
this makes character string cvecs (and ngrams) usable for image indexing.

Minor unrelated change - removed the check for puzzle.h. In theory it is
nice to have; but it was causing issues on my dev machine, FreeBSD 10.3.

Added 3 utlities, but they are not installed via CPAN as of yet. They
mainly exist currently to compare images via commandline using the bindings
provided by this library.